LayoutTests/html5lib/generated/run-tests1-data-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
../resources/tests1.dat: PASS
